pyproject.toml
[project]
name = "hybrid-interpretability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.50.0",
    "dotenv>=0.9.9",
    "jupyter>=1.1.1",
    "kagglehub>=0.3.12",
    "matplotlib>=3.10.1",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "recurrentgemma",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "tiktoken>=0.9.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]

[tool.uv.sources]
recurrentgemma = { git = "https://github.com/stevenabreu7/recurrentgemma.git", rev = "dejavu" }
